fixed the last of the ads logic and implemented restrictive logic

+++ JensenAuktion/Controllers/AdsController.cs
@@ -1,5 +1,7 @@
-﻿using JensenAuktion.Repository.Entities;
+﻿using JensenAuktion.Repository.DTO;
+using JensenAuktion.Repository.Entities;
 using JensenAuktion.Repository.Interfaces;
+using JensenAuktion.Services;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
@@ -10,26 +12,14 @@ namespace JensenAuktion.Controllers
     public class AdsController : ControllerBase
     {
         private readonly IAdRepository _adRepository;
-        public AdsController(IAdRepository adRepository)
+        private readonly IAdsService _adsService;
+        public AdsController(IAdRepository adRepository, IAdsService adsService)
         {
             _adRepository = adRepository;
+            _adsService = adsService;
         }
 
-        [HttpPost]
-        //[Authorize]
-        public IActionResult CreateAd(Ad ad)
-        {
-            try
-            {
-                _adRepository.CreateAd(ad);
-            }
-            catch (Exception ex)
-            {
-                return BadRequest(new { message = ex.Message });
-            }
 
-            return Ok(ad);
-        }
 
         [HttpGet]
         public IActionResult GetAllAds()
@@ -45,12 +35,57 @@ namespace JensenAuktion.Controllers
             }
         }
 
+        [HttpPost]
+        [Authorize]
+        public IActionResult CreateAd(AdsCreateDTO ad)
+        {
+            try
+            {
+                _adRepository.CreateAd(ad);
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(new { message = ex.Message });
+            }
+
+            return Ok(ad);
+        }
+
+        [HttpGet("{id}")]
+        public IActionResult GetAdById(int id)
+        {
+            try
+            {
+                if (_adsService.IsAdClosed(id))
+                {
+
+                    return Ok(_adRepository.GetClosedAdById(id));
+
+                }
+                else
+                {
+                    return Ok(_adRepository.GetAdById(id));
+                }
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(ex.Message);
+            }
+
+        }
+
         [HttpPut]
-        //[Authorize]
+        [Authorize]
         public IActionResult UpdateAd(Ad ad)
         {
             try
             {
+                if (_adsService.CheckIfAdHasBids(ad.AdID))
+                {
+                    _adRepository.UpdateAdWithBid(ad);
+                    return Ok("Updated without price due to bids " + ad);
+                }
+
                 _adRepository.UpdateAd(ad);
             }
             catch (Exception ex)
@@ -63,11 +98,16 @@ namespace JensenAuktion.Controllers
 
 
         [HttpDelete("{id}")]
-        //[Authorize]
+        [Authorize]
         public IActionResult DeleteAd(int id)
         {
             try
             {
+                if (_adsService.CheckIfAdHasBids(id))
+                {
+                    return BadRequest(new { message = "Cannot delete ad with existing bids" });
+                }
+
                 _adRepository.DeleteAd(id);
             }
             catch (Exception ex)
