+++ JensenAuktion/Controllers/BidController.cs
@@ -1,8 +1,7 @@
 ﻿using JensenAuktion.Repository.Entities;
 using JensenAuktion.Repository.Interfaces;
-using JensenAuktion.Repository.Repos;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 
 namespace JensenAuktion.Controllers
 {
@@ -11,23 +10,31 @@ namespace JensenAuktion.Controllers
     public class BidController : ControllerBase
     {
         private readonly IBidRepo _bidRepo;
+        private readonly IAdsService _adsService;
 
-        public BidController(IBidRepo bidRepo)
+        public BidController(IBidRepo bidRepo, IAdsService adsService)
         {
             _bidRepo = bidRepo;
+            _adsService = adsService;
         }
 
         // POST: api/Bid
         [HttpPost]
         public IActionResult CreateBid([FromBody] Bid bid)
         {
-            if (bid == null)
-            {
-                return BadRequest("Bid data is invalid.");
-            }
-
+            if (bid == null) return BadRequest("Bid data is invalid.");
+            var userIdFromToken = User.FindFirstValue(ClaimTypes.Sid);
+            var userId = _adsService.CheckAdUserId(bid.AdID).ToString();
             try
             {
+                if (_adsService.IsAdClosed(bid.AdID))
+                {
+                    return BadRequest("You can't place a bid on a closed Ad");
+                }
+                else if (userId == userIdFromToken)
+                {
+                    return BadRequest("You can't place a bid on your own ad");
+                }
                 int newBidID = _bidRepo.CreateBid(bid);
                 return Ok(new { BidID = newBidID, Message = "Bid created successfully." });
             }
@@ -37,26 +44,23 @@ namespace JensenAuktion.Controllers
             }
         }
 
-        
         [HttpDelete("{id}")]
         public IActionResult DeleteBid(int id)
         {
             try
             {
+                if (_adsService.IsAdClosedByBid(id))
+                {
+                    return BadRequest("You can't remove a bid on a closed Ad");
+                }
+
                 bool isDeleted = _bidRepo.DeleteBid(id);
-                if (isDeleted)
-                {
-                    return Ok(new { Message = "Bid deleted successfully." });
-                }
-                else
-                {
-                    return NotFound($"Bid with ID {id} not found.");
-                }
             }
             catch (Exception ex)
             {
                 return StatusCode(500, $"Internal server error: {ex.Message}");
             }
+            return Ok(id + " Is deleted");
         }
     }
 }
